Proxy Topics: handling background sessions (part 1).
Background sessions are handled by the proxy topics, not master topics.
We add such sessions to the deferred notifications list at the time
when the response to the subscription request from the master is received
and count the ticks on such sessions from that point.

When a session becomes due, we poll it off the notifications list and
send the notifications request to the master (this part is a TODO -
will come in a follow up PR).