Use AND of ORs for required terms in search queries.
When we rewrite tokens in search queries, we want to be able
to search for both the original and the rewritten terms.
Example:
- user `alice` has a tag `basic:alice`
- group `travel` is tagged `travel`

When we search for `travel` we would like to retrieve and return both results.

In order to achieve this, change the search interface as follows by
making required terms as an AND of ORs.